+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
-- Added console command to list all aliases and indexes of the configured Elasticsearch cluster. 
+- Added console command to list all aliases and indexes of the configured Elasticsearch cluster.
+- Added workflow to switch back to Craft's database index.
+- Added console command to truncate Craft's database index table.
 
 ### Fixed